backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
sqlalchemy==2.0.29
aiosqlite==0.20.0
python-multipart==0.0.9
httpx==0.27.0
firebase-admin==6.5.0
websockets==12.0
